split container-diff into subcommands: analyze and diff
