+++ cmd/root.go
@@ -8,9 +8,7 @@ import (
 	"os"
 	"reflect"
 	"sort"
-	"sync"
 
-	"github.com/GoogleCloudPlatform/container-diff/differs"
 	"github.com/GoogleCloudPlatform/container-diff/utils"
 	"github.com/golang/glog"
 	"github.com/spf13/cobra"
@@ -20,7 +18,6 @@ import (
 var json bool
 var eng bool
 var save bool
-
 var apt bool
 var node bool
 var file bool
@@ -35,138 +32,12 @@ var analyzeFlagMap = map[string]*bool{
 	"pip":     &pip,
 }
 
+type validatefxn func(args []string) (bool, error)
+
 var RootCmd = &cobra.Command{
 	Use:   "To analyze a single image: [image].  To compare two images: [image1] [image2]",
 	Short: "Analyze a single image or compare two images.",
 	Long:  `Analyzes a single image or compares two images using the specifed analyzers/differs as indicated via flags (see documentation for available ones).`,
-	Run: func(cmd *cobra.Command, args []string) {
-		if validArgs, err := validateArgs(args); !validArgs {
-			glog.Error(err.Error())
-			os.Exit(1)
-		}
-
-		utils.SetDockerEngine(eng)
-
-		analyzeArgs := []string{}
-		allAnalyzers := getAllAnalyzers()
-		for _, name := range allAnalyzers {
-			if *analyzeFlagMap[name] == true {
-				analyzeArgs = append(analyzeArgs, name)
-			}
-		}
-
-		// If no differs/analyzers are specified, perform them all as the default
-		if len(analyzeArgs) == 0 {
-			analyzeArgs = allAnalyzers
-		}
-
-		var err error
-		// In the case of one image, "analyzes" it
-		// In the case of two images, takes their "diff"
-		if len(args) == 1 {
-			err = analyzeImage(args[0], analyzeArgs)
-		} else {
-			err = diffImages(args[0], args[1], analyzeArgs)
-		}
-
-		if err != nil {
-			glog.Error(err)
-			os.Exit(1)
-		}
-	},
-}
-
-func diffImages(image1Arg, image2Arg string, diffArgs []string) error {
-	var wg sync.WaitGroup
-	wg.Add(2)
-
-	glog.Infof("Starting diff on images %s and %s, using differs: %s", image1Arg, image2Arg, diffArgs)
-
-	var image1, image2 utils.Image
-	var err error
-	go func() {
-		defer wg.Done()
-		image1, err = utils.ImagePrepper{image1Arg}.GetImage()
-		if err != nil {
-			glog.Error(err.Error())
-		}
-	}()
-
-	go func() {
-		defer wg.Done()
-		image2, err = utils.ImagePrepper{image2Arg}.GetImage()
-		if err != nil {
-			glog.Error(err.Error())
-		}
-	}()
-	wg.Wait()
-	if err != nil {
-		cleanupImage(image1)
-		cleanupImage(image2)
-		return errors.New("Could not perform image diff")
-	}
-
-	diffTypes, err := differs.GetAnalyzers(diffArgs)
-	if err != nil {
-		glog.Error(err.Error())
-		cleanupImage(image1)
-		cleanupImage(image2)
-		return errors.New("Could not perform image diff")
-	}
-
-	req := differs.DiffRequest{image1, image2, diffTypes}
-	if diffs, err := req.GetDiff(); err == nil {
-		glog.Info("Retrieving diffs")
-		outputResults(diffs)
-		if !save {
-			cleanupImage(image1)
-			cleanupImage(image2)
-
-		} else {
-			dir, _ := os.Getwd()
-			glog.Infof("Images were saved at %s as %s and %s", dir, image1.FSPath, image2.FSPath)
-		}
-	} else {
-		glog.Error(err.Error())
-		cleanupImage(image1)
-		cleanupImage(image2)
-		return errors.New("Could not perform image diff")
-	}
-
-	return nil
-}
-
-func analyzeImage(imageArg string, analyzerArgs []string) error {
-	image, err := utils.ImagePrepper{imageArg}.GetImage()
-	if err != nil {
-		glog.Error(err.Error())
-		cleanupImage(image)
-		return errors.New("Could not perform image analysis")
-	}
-	analyzeTypes, err := differs.GetAnalyzers(analyzerArgs)
-	if err != nil {
-		glog.Error(err.Error())
-		cleanupImage(image)
-		return errors.New("Could not perform image analysis")
-	}
-
-	req := differs.SingleRequest{image, analyzeTypes}
-	if analyses, err := req.GetAnalysis(); err == nil {
-		glog.Info("Retrieving analyses")
-		outputResults(analyses)
-		if !save {
-			cleanupImage(image)
-		} else {
-			dir, _ := os.Getwd()
-			glog.Infof("Image was saved at %s as %s", dir, image.FSPath)
-		}
-	} else {
-		glog.Error(err.Error())
-		cleanupImage(image)
-		return errors.New("Could not perform image analysis")
-	}
-
-	return nil
 }
 
 func outputResults(resultMap map[string]utils.Result) {
@@ -215,33 +86,16 @@ func getAllAnalyzers() []string {
 	return allAnalyzers
 }
 
-func validateArgs(args []string) (bool, error) {
-	validArgNum, err := checkArgNum(args)
-	if err != nil {
-		return false, err
-	} else if !validArgNum {
-		return false, nil
-	}
-	validArgType, err := checkArgType(args)
-	if err != nil {
-		return false, err
-	} else if !validArgType {
-		return false, nil
+func validateArgs(args []string, validatefxns ...validatefxn) (bool, error) {
+	for _, validatefxn := range validatefxns {
+		valid, err := validatefxn(args)
+		if err != nil {
+			return false, err
+		} else if !valid {
+			return false, nil
+		}
 	}
 	return true, nil
-}
-
-func checkArgNum(args []string) (bool, error) {
-	var errMessage string
-	if len(args) < 1 {
-		errMessage = "Too few arguments. Should have one or two images as arguments."
-		return false, errors.New(errMessage)
-	} else if len(args) > 2 {
-		errMessage = "Too many arguments. Should have at most two images as arguments."
-		return false, errors.New(errMessage)
-	} else {
-		return true, nil
-	}
 }
 
 func checkImage(arg string) bool {
@@ -287,13 +141,16 @@ func remove(path string, dir bool) string {
 
 func init() {
 	pflag.CommandLine.AddGoFlagSet(goflag.CommandLine)
-	RootCmd.Flags().BoolVarP(&json, "json", "j", false, "JSON Output defines if the diff should be returned in a human readable format (false) or a JSON (true).")
-	RootCmd.Flags().BoolVarP(&eng, "eng", "e", false, "By default the docker calls are shelled out locally, set this flag to use the Docker Engine Client (version compatibility required).")
-	RootCmd.Flags().BoolVarP(&pip, "pip", "p", false, "Set this flag to use the pip differ.")
-	RootCmd.Flags().BoolVarP(&node, "node", "n", false, "Set this flag to use the node differ.")
-	RootCmd.Flags().BoolVarP(&apt, "apt", "a", false, "Set this flag to use the apt differ.")
-	RootCmd.Flags().BoolVarP(&file, "file", "f", false, "Set this flag to use the file differ.")
-	RootCmd.Flags().BoolVarP(&history, "history", "d", false, "Set this flag to use the dockerfile history differ.")
-	RootCmd.Flags().BoolVarP(&save, "save", "s", false, "Set this flag to save rather than remove the final image filesystems on exit.")
-	RootCmd.Flags().BoolVarP(&utils.SortSize, "order", "o", false, "Set this flag to sort any file/package results by descending size. Otherwise, they will be sorted by name.")
+}
+
+func addSharedFlags(cmd *cobra.Command) {
+	cmd.Flags().BoolVarP(&json, "json", "j", false, "JSON Output defines if the diff should be returned in a human readable format (false) or a JSON (true).")
+	cmd.Flags().BoolVarP(&eng, "eng", "e", false, "By default the docker calls are shelled out locally, set this flag to use the Docker Engine Client (version compatibility required).")
+	cmd.Flags().BoolVarP(&pip, "pip", "p", false, "Set this flag to use the pip differ.")
+	cmd.Flags().BoolVarP(&node, "node", "n", false, "Set this flag to use the node differ.")
+	cmd.Flags().BoolVarP(&apt, "apt", "a", false, "Set this flag to use the apt differ.")
+	cmd.Flags().BoolVarP(&file, "file", "f", false, "Set this flag to use the file differ.")
+	cmd.Flags().BoolVarP(&history, "history", "d", false, "Set this flag to use the dockerfile history differ.")
+	cmd.Flags().BoolVarP(&save, "save", "s", false, "Set this flag to save rather than remove the final image filesystems on exit.")
+	cmd.Flags().BoolVarP(&utils.SortSize, "order", "o", false, "Set this flag to sort any file/package results by descending size. Otherwise, they will be sorted by name.")
 }
